Move health checks row scheme to schemes folder
