+++ src/js/components/HealthChecksComponent.jsx
@@ -5,7 +5,8 @@ var DuplicableRowsMixin = require("../mixins/DuplicableRowsMixin");
 var FormGroupComponent = require("../components/FormGroupComponent");
 var HealthCheckProtocols = require("../constants/HealthCheckProtocols");
 
-const healthChecksRowScheme = require("../stores/healthChecksRowScheme");
+const healthChecksRowScheme =
+  require("../stores/schemes/healthChecksRowScheme");
 
 const numberInputAttributes = {
   min: 0,
